requirements.txt
Flask==3.0.1
flask-login==0.6.3
pymongo==4.6.1
dnspython==2.6.1
python-dotenv==1.0.1
Werkzeug==3.0.1

